+++ packages/plugins/IconMenu/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unpublished
 
+- Feature: Add locale for new plugin `@polar/plugin-routing`.
 - Fix: Adjust behaviour of setting content to the MoveHandle for small devices so that an already opened plugin is still open on resize.
 
 ## 1.4.0